+++ src/main/java/io/github/mortuusars/chalk/client/gui/ChalkBoxScreen.java
@@ -1,16 +1,19 @@
 package io.github.mortuusars.chalk.client.gui;
 
+import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import io.github.mortuusars.chalk.Chalk;
-import io.github.mortuusars.chalk.config.Config;
+import io.github.mortuusars.chalk.Config;
 import io.github.mortuusars.chalk.item.ChalkBoxItem;
 import io.github.mortuusars.chalk.menus.ChalkBoxMenu;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.Slot;
 import org.jetbrains.annotations.NotNull;
 
@@ -18,12 +21,13 @@ public class ChalkBoxScreen extends AbstractContainerScreen<ChalkBoxMenu> {
     public static final ResourceLocation TEXTURE = Chalk.resource("textures/gui/container/chalk_box.png");
     private static final int GLOWING_BAR_WIDTH = 72;
     private final int maxGlowingUses;
+    private final Player player;
 
     public ChalkBoxScreen(ChalkBoxMenu menu, Inventory playerInventory, Component title) {
         super(menu, playerInventory, title);
         maxGlowingUses = Config.Common.CHALK_BOX_GLOWING_AMOUNT_PER_ITEM.get();
-
         this.minecraft = Minecraft.getInstance();
+        this.player = Minecraft.getInstance().player;
     }
 
     @Override
@@ -43,17 +47,17 @@ public class ChalkBoxScreen extends AbstractContainerScreen<ChalkBoxMenu> {
     public void render(@NotNull GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
         super.render(graphics, mouseX, mouseY, partialTick);
 
-        if (getMenu().chalkBoxCoords != null) {
-            graphics.renderItem(getMenu().getChalkBoxStack(), getGuiLeft() + getMenu().chalkBoxCoords.getFirst(), getGuiTop() + getMenu().chalkBoxCoords.getSecond());
+        for (Slot slot : getMenu().slots) {
+            if (!slot.mayPickup(player) && !slot.isActive()) {
+                RenderSystem.enableBlend();
+                RenderSystem.defaultBlendFunc();
+                graphics.blit(TEXTURE, leftPos + slot.x - 1, topPos + slot.y - 1, 176, 36, 20, 20);
+                graphics.renderFakeItem(slot.getItem(), leftPos + slot.x, topPos + slot.y);
+                graphics.fill(RenderType.guiGhostRecipeOverlay(), leftPos + slot.x - 1, topPos + slot.y - 1,
+                        leftPos + slot.x + 15, topPos + slot.y + 15, 0x40FFFFFF);
+                RenderSystem.disableBlend();
+            }
         }
-
-        PoseStack poseStack = graphics.pose();
-        poseStack.pushPose();
-        poseStack.translate(0, 0, 200);
-        graphics.fill(getGuiLeft() + getMenu().chalkBoxCoords.getFirst() - 1, getGuiTop() + getMenu().chalkBoxCoords.getSecond() - 1,
-                getGuiLeft() + getMenu().chalkBoxCoords.getFirst() + 17, getGuiTop() + getMenu().chalkBoxCoords.getSecond() + 17,
-                0x20c8c8c8);
-        poseStack.popPose();
 
         this.renderTooltip(graphics, mouseX, mouseY);
     }
@@ -62,17 +66,9 @@ public class ChalkBoxScreen extends AbstractContainerScreen<ChalkBoxMenu> {
     protected void renderBg(GuiGraphics graphics, float partialTick, int mouseX, int mouseY) {
         graphics.blit(TEXTURE, getGuiLeft(), getGuiTop(), 0, 0, imageWidth, imageHeight);
 
+        renderChalkSlots(graphics, 52, getMenu().isGlowingEnabled() ? 17 : 32);
+
         if (getMenu().isGlowingEnabled()) {
-            // Chalk Slots
-            graphics.blit(TEXTURE, getGuiLeft() + 52, getGuiTop() + 17, 0, 180, 72, 36);
-
-            for (int slotIndex = 0; slotIndex < ChalkBoxItem.CHALK_SLOTS; slotIndex++) {
-                Slot slot = getMenu().slots.get(slotIndex);
-                if (slot.getItem().isEmpty()) {
-                    graphics.blit(TEXTURE, getGuiLeft() + slot.x - 1, getGuiTop() + slot.y - 1, 176, 0, 18, 18);
-                }
-            }
-
             // Bar + Slot
             graphics.blit(TEXTURE, getGuiLeft() + 52, getGuiTop() + 57, 0, 217, 72, 28);
 
@@ -85,14 +81,17 @@ public class ChalkBoxScreen extends AbstractContainerScreen<ChalkBoxMenu> {
             int barSize = (int) Math.ceil((Math.min(getMenu().getGlowAmount(), maxGlowingUses) / (float) maxGlowingUses) * GLOWING_BAR_WIDTH);
             int glowingBarFillLevel = Math.min(GLOWING_BAR_WIDTH, barSize);
             graphics.blit(TEXTURE, getGuiLeft() + 52, getGuiTop() + 57, 72, 217, glowingBarFillLevel, 5);
-
-        } else {
-            // Chalk slots
-            graphics.blit(TEXTURE, getGuiLeft() + 52, getGuiTop() + 32, 0, 180, 72, 36);
         }
+    }
 
-        graphics.fill(getGuiLeft() + getMenu().chalkBoxCoords.getFirst() - 1, getGuiTop() + getMenu().chalkBoxCoords.getSecond() - 1,
-                getGuiLeft() + getMenu().chalkBoxCoords.getFirst() + 17, getGuiTop() + getMenu().chalkBoxCoords.getSecond() + 17,
-                0xAAc8c8c8);
+    protected void renderChalkSlots(GuiGraphics graphics, int x, int y) {
+        graphics.blit(TEXTURE, getGuiLeft() + x, getGuiTop() + y, 0, 180, 72, 36);
+
+        for (int slotIndex = 0; slotIndex < ChalkBoxItem.CHALK_SLOTS; slotIndex++) {
+            Slot slot = getMenu().slots.get(slotIndex);
+            if (slot.getItem().isEmpty()) {
+                graphics.blit(TEXTURE, getGuiLeft() + slot.x - 1, getGuiTop() + slot.y - 1, 176, 0, 18, 18);
+            }
+        }
     }
 }