+++ src/main/java/io/github/mortuusars/chalk/core/SymbolUnlocking.java
@@ -2,7 +2,7 @@ package io.github.mortuusars.chalk.core;
 
 import com.mojang.datafixers.util.Pair;
 import io.github.mortuusars.chalk.Chalk;
-import io.github.mortuusars.chalk.config.Config;
+import io.github.mortuusars.chalk.Config;
 import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.AdvancementProgress;
 import net.minecraft.resources.ResourceLocation;